+++ src/plugins/normalState/edit/editNormalState.component.ts
@@ -4,14 +4,14 @@ import {StringLocalization, STRING_LOCALIZATION} from '@anglr/common';
 import {Subscription} from 'rxjs';
 
 import {EditNormalStateOptions, EditNormalState} from './editNormalState.interface';
-import {NgSelectPlugin, TemplateGatherer} from '../../../misc';
+import {NgSelectPlugin} from '../../../misc';
 import {NgSelectPluginInstances} from '../../../components/select';
 import {NG_SELECT_PLUGIN_INSTANCES} from '../../../components/select/types';
 import {NormalStateTexts} from '../normalState.interface';
 import {NORMAL_STATE_OPTIONS} from '../types';
 import {ValueHandler} from '../../valueHandler';
 import {VALUE_HANDLER} from '../../valueHandler/types';
-import {NgSelectOption} from '../../../components/option';
+import {PluginBus} from '../../../misc/pluginBus/pluginBus';
 
 /**
  * Default options for normal state
@@ -29,8 +29,7 @@ const defaultOptions: EditNormalStateOptions =
     texts:
     {
         nothingSelected: 'Nothing selected'
-    },
-    readonly: false
+    }
 };
 
 /**
@@ -76,26 +75,6 @@ export class EditNormalStateComponent implements EditNormalState, NgSelectPlugin
         this._options = extend(true, this._options, options);
     }
 
-    /**
-     * Occurs when user clicks on normal state
-     */
-    public click: EventEmitter<void> = new EventEmitter<void>();
-
-    /**
-     * Occurs when normal state gains focus
-     */
-    public focus: EventEmitter<void> = new EventEmitter<void>();
-
-    /**
-     * Occurs when user tries to cancel one of selected values
-     */
-    public cancelOption: EventEmitter<NgSelectOption> = new EventEmitter<NgSelectOption>();
-
-    /**
-     * Gatherer used for obtaining custom templates
-     */
-    public templateGatherer: TemplateGatherer;
-
     //######################### public properties - template bindings #########################
 
     /**
@@ -112,6 +91,7 @@ export class EditNormalStateComponent implements EditNormalState, NgSelectPlugin
 
     //######################### constructor #########################
     constructor(@Inject(NG_SELECT_PLUGIN_INSTANCES) @Optional() public ngSelectPlugins: NgSelectPluginInstances,
+                @Optional() public pluginBus: PluginBus,
                 public pluginElement: ElementRef,
                 protected _changeDetector: ChangeDetectorRef,
                 @Inject(STRING_LOCALIZATION) protected _stringLocalization: StringLocalization,
@@ -129,11 +109,8 @@ export class EditNormalStateComponent implements EditNormalState, NgSelectPlugin
     {
         this._destroyed = true;
 
-        if(this._textsChangedSubscription)
-        {
-            this._textsChangedSubscription.unsubscribe();
-            this._textsChangedSubscription = null;
-        }
+        this._textsChangedSubscription?.unsubscribe();
+        this._textsChangedSubscription = null;
     }
 
     //######################### public methods - implementation of EditNormalState #########################